requirements.txt
praw
python-dotenv
tqdm
requests
